% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ipcc.R
\name{.ipcc_fetch_term_index}
\alias{.ipcc_fetch_term_index}
\title{Collect the IPCC term index}
\usage{
.ipcc_fetch_term_index(pause = 0.3)
}
\value{
A data frame with \code{id}, \code{term} and \code{prefix}.
}
\description{
Collect the IPCC term index
}
\keyword{internal}
